[RESOURCES] moved external resources, changed background handling a bit
