+++ resources/backgrounds/mapping.yaml
@@ -1,0 +1,8 @@
+header.overview: awesomebackdrop_wide.webp
+header.extensions: null
+header.config: null
+header.launch: last_tab.webp
+tab.overview: bg_richard.webp
+tab.extensions: bg_richard.webp
+tab.config: bg_richard.webp
+tab.launch: bg_richard.webp
