+++ resources/gameinfo/game-version.yaml
