+++ resources/lang/sources/de/gui-editor.yaml
