+++ resources/lang/sources/de/gui-general.yaml
