+++ resources/lang/sources/de/gui-landing.yaml
