+++ resources/lang/sources/de/gui-launch.yaml
