+++ resources/lang/sources/en/gui-editor.yaml
