+++ resources/lang/sources/en/gui-general.yaml
