+++ resources/lang/sources/en/gui-landing.yaml
