+++ resources/lang/sources/en/gui-launch.yaml
