--- src-tauri/backgrounds/mapping.yaml
@@ -1,6 +1,0 @@
-header: bg_richard.webp
-tabs:
-  overview: bg_richard.webp
-  extensions: null
-  config: null
-  launch: null
