--- src/components/general/gradient-img.css
@@ -1,15 +1,0 @@
-.gradient-img {
-  position: absolute;
-  height: 100%;
-  right: 0;
-  top: 0;
-  -webkit-mask-image: linear-gradient(100deg, transparent 20%, black 70%);
-  mask-image: linear-gradient(100deg, transparent 20%, black 70%);
-  z-index: -1;
-}
-
-.gradient-img--width {
-  height: 100%;
-  width: 50%;
-  object-fit: cover;
-}
