+++ src/components/ucp-tabs/common/gradient-img/gradient-img.css
@@ -1,0 +1,20 @@
+.gradient-img {
+  position: absolute;
+  height: 100%;
+  top: 0;
+  z-index: -1;
+}
+
+.gradient-img--header {
+  left: 0;
+  width: 100%;
+  object-fit: cover;
+  -webkit-mask-image: linear-gradient(0deg, transparent, black 30%);
+  mask-image: linear-gradient(0deg, rgb(0 0 0 / 20%), black 30%);
+}
+
+.gradient-img--tab {
+  right: 0;
+  -webkit-mask-image: linear-gradient(100deg, transparent 20%, black 70%);
+  mask-image: linear-gradient(100deg, transparent 20%, black 70%);
+}
