+++ src/components/ucp-tabs/launch/launch.css
@@ -6,9 +6,10 @@
   display: flex;
   flex-direction: row;
   width: 100%;
-  padding-top: 15px;
-}
+  padding: 15px 0 15px 0;
 
+  background-color: var(--ucp-menu-shade);
+}
 .launch__options {
   align-items: start;
   padding: 15px;
