+++ src/components/ucp-tabs/ucp-tabs.tsx
@@ -26,7 +26,7 @@ import Overview from './overview/overview';
 import Launch from './launch/launch';
 import { CURRENT_DISPLAYED_TAB, UITabs } from './tabs-state';
 import ExtensionManager from './extension-manager/extension-manager';
-import GradientImg from '../general/gradient-img';
+import GradientImg from './common/gradient-img/gradient-img';
 import {
   loadYaml,
   receiveAssetUrl,
@@ -39,47 +39,37 @@ import {
 
 const LOGGER = new Logger('ucp-taps.tsx');
 
-interface BackgroundMapping {
-  header?: string;
-  tabs?: Record<UITabs, string | undefined>;
-}
-
 const DISPLAY_CONFIG_TABS_ATOM = atom(
   (get) => get(INIT_DONE) && !get(INIT_RUNNING) && !get(INIT_ERROR),
 );
 
-const BACKGROUNDS_PATH_ATOM: Atom<Promise<BackgroundMapping>> = atom(
-  async () => {
-    const mapping: BackgroundMapping = await resolveResourcePath([
-      BACKGROUNDS_DIRECTORY,
-      BACKGROUNDS_MAPPING_FILE,
-    ])
-      .then(loadYaml)
-      .then((res) => res.getOrThrow())
-      .catch((err) => {
-        LOGGER.msg('Failed to load background mappings file: {}', err).error();
-        return {};
-      });
+const BACKGROUNDS_PATH_ATOM: Atom<Promise<Record<string, string | undefined>>> =
+  atom(async () => {
+    const mapping: Record<string, string | undefined> =
+      await resolveResourcePath([
+        BACKGROUNDS_DIRECTORY,
+        BACKGROUNDS_MAPPING_FILE,
+      ])
+        .then(loadYaml)
+        .then((res) => res.getOrThrow())
+        .catch((err) => {
+          LOGGER.msg(
+            'Failed to load background mappings file: {}',
+            err,
+          ).error();
+          return {};
+        });
 
     // resolve paths on load
-    mapping.header = !mapping.header
-      ? undefined
-      : await resolveResourcePath([BACKGROUNDS_DIRECTORY, mapping.header])
-          .then(receiveAssetUrl)
-          .catch(() => undefined);
-    if (!mapping.tabs) {
-      return mapping;
-    }
-    for (const [tab, file] of Object.entries(mapping.tabs)) {
-      mapping.tabs[tab as UITabs] = !file
+    for (const [key, file] of Object.entries(mapping)) {
+      mapping[key] = !file
         ? undefined
         : await resolveResourcePath([BACKGROUNDS_DIRECTORY, file])
             .then(receiveAssetUrl)
             .catch(() => undefined);
     }
     return mapping;
-  },
-);
+  });
 
 export default function UcpTabs() {
   const { t } = useTranslation(['gui-general', 'gui-editor', 'gui-launch']);
@@ -103,8 +93,8 @@ export default function UcpTabs() {
   const [currentTab, setCurrentTab] = useAtom(CURRENT_DISPLAYED_TAB);
 
   const backgroundMapping = useAtomValue(BACKGROUNDS_PATH_ATOM);
-  const headerImage = backgroundMapping?.header ?? '';
-  const tabImage = backgroundMapping?.tabs?.[currentTab] ?? '';
+  const headerImage = backgroundMapping?.[`header.${currentTab}`] ?? '';
+  const tabImage = backgroundMapping?.[`tab.${currentTab}`] ?? '';
 
   return (
     <div
@@ -117,7 +107,7 @@ export default function UcpTabs() {
         onSelect={(newKey) => setCurrentTab(newKey as UITabs)}
       >
         <Nav variant="tabs" className="ucp-tabs-header" data-tauri-drag-region>
-          <GradientImg src={headerImage} basedOnWidth />
+          <GradientImg src={headerImage} type="header" />
 
           <Nav.Item>
             <Nav.Link
@@ -186,7 +176,7 @@ export default function UcpTabs() {
           </Nav.Item>
         </Nav>
         <Tab.Content className="ornament-border">
-          <GradientImg src={tabImage} />
+          <GradientImg src={tabImage} type="tab" />
 
           <Tab.Pane eventKey="overview" className="tab-panel">
             <Overview />
